Fix product listing page (#24)
* Change intro slides content

* Minor tweaks to wording

* Re-point Keycloak govuk theme submodule commit

* Add banner to non-opss page (#1340)

* Install monitoring with Scout (#1341)

* Install monitoring with Scout

* Appease rubocop

* Comment out randomly failing tests

* Remove rack-timeout

* Add gemfile lock for cosmetics-web and psd-web

* Rm gemfile.lock from shared-web

* PSD Date received bug fix (#1339)

* Stopped date component being required, which was checking on update. Added date_received blank validation to the enquiry model.

* Updated changelog

* Use normal validation, but prevent if partially filled

* Add model tests to cover the bug fix

* Rubocop copping

* PSD: Only show enquiry fields on enquiry pages

* Add enquiry? method on investigation model

* Added model tests

* Rubocopped

* Update product categories (#1338)

* PSD: Re-add google analytics after being put in a lost content_for (#1355)

* PSD: Various bug fixes (#1219)

* PSD-1157 Fixing text overlapping images

* PSD-1198 Duplicate header on TS flow

* PSD-1180 Decapitalize emails

* PSD-913 Autocomplete has a blank entry

* Why_reporting missing page title

* Better error messages for product category and hazard type, using en file to give custom error messages

* Cookie banner crashing

* Update PaaS production password

* Change to force prod deploy

* Fix product page

1. make countrie not read from the cache - Just hold it in memory
2. fix rails cache store config

* appease rubcop
