+++ psd-web/config/environments/production.rb
@@ -49,10 +49,8 @@ Rails.application.configure do
   config.log_tags = [:request_id]
 
   # Use a different cache store in production.
-  config.cache_store = :memory_store
-  config.cache_store = :redis_cache_store, {
-    url: Rails.application.config_for(:redis_store)
-  }
+  config.cache_store = :redis_cache_store, Rails.application.config_for(:redis_store).symbolize_keys
+
   config.action_mailer.perform_caching = false
 
   # Url for mailer
